-auth -catalog -logs -navstevy pacientu (auth:admin,pacient) -fixes

+++ src/main/java/com/clinicappbdas2/controller/UsersController.java
@@ -31,7 +31,7 @@ public class UsersController {
 //    }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getById(@PathVariable Integer id) {
+    public ResponseEntity<User> getById(@PathVariable Long id) {
         return ResponseEntity.ok(userService.getById(id));
     }
